+++ plugins/module_utils/api_endpoints.py
@@ -1,3 +1,12 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+#
+# Apache License v2.0+ (see LICENSE or https://www.apache.org/licenses/LICENSE-2.0)
+
+from __future__ import absolute_import, division, print_function
+
+__metaclass__ = type
+
 from enum import Enum
 
 
@@ -12,3 +21,5 @@ class CDOAPI(Enum):
     ASA_CONFIG = f"{SVCS}/asa/configs"
     SPECIFIC_DEVICE = "aegis/rest/v1/device/{uid}/specific-device"
     FMC_ACCESS_POLICY = "fmc/api/fmc_config/v1/domain/{domain_uid}/policy/accesspolicies"
+    DEPLOY = "aegis/rest/v1/services/targets/device-changelog"
+    JOBS = "aegis/rest/v1/services/state-machines/jobs"